+++ pyslvs_ui/qt_patch.py
@@ -5,25 +5,27 @@
 Customized class will define below.
 """
 
-__all__ = ['API', 'QT_VERSION', 'qt_image_suffix', 'qt_image_format', 'QABCMeta']
+__all__ = ['API', 'QT_VERSION', 'qt_image_suffix', 'qt_image_format',
+           'QABCMeta']
 __author__ = "Yuan Chang"
 __copyright__ = "Copyright (C) 2016-2020"
 __license__ = "AGPL"
 __email__ = "pyslvs@gmail.com"
 
-from typing import Any
 from abc import ABCMeta
-from qtpy import QtCore, API_NAME
+from qtpy import API_NAME
+from qtpy.QtCore import PYQT_VERSION_STR, __version__, QEvent, QObject
+from qtpy.QtWidgets import QApplication
 from importlib import import_module
 
 if API_NAME == 'PyQt5':
-    API = f"{API_NAME} {QtCore.PYQT_VERSION_STR}"
+    API = f"{API_NAME} {PYQT_VERSION_STR}"
 elif API_NAME == 'PySide2':
     API = f"{API_NAME} {getattr(import_module('PySide2'), '__version__')}"
 else:
     raise ModuleNotFoundError("module not found: PyQt5 or PySide2")
 
-QT_VERSION = QtCore.__version__
+QT_VERSION = __version__
 qt_image_suffix = []
 qt_image_format = []
 for suffix, name in (
@@ -43,15 +45,16 @@ qt_image_suffix = tuple(qt_image_suffix)
 qt_image_format = tuple(qt_image_format)
 del suffix, name
 
-_Wrapper: Any = type(QtCore.QObject)
+# Avoid Qt clipboard loss data after program closed
+QApplication.sendEvent(QApplication.clipboard(), QEvent(QEvent.Clipboard))
 
 
-class QABCMeta(_Wrapper, ABCMeta):
+class QABCMeta(type(QObject), ABCMeta):  # type: ignore
     """Qt ABCMeta class.
 
     Usage:
     >>> from abc import abstractmethod
-    >>> class MyQObject(QtCore.QObject, metaclass=QABCMeta):
+    >>> class MyQObject(QObject, metaclass=QABCMeta):
     >>>    @abstractmethod
     >>>    def my_abstract_method(self) -> None:
     >>>        ...